% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MriCloudR.r
\docType{methods}
\name{downloadResult}
\alias{downloadResult}
\alias{downloadResult,MriCloudR,character,character,logical-method}
\alias{downloadResult,MriCloudR,character,missing,logical-method}
\title{Download result of processing request}
\usage{
downloadResult(object, jobId = "character", filename = "character",
  waitForJobToFinish = TRUE)

\S4method{downloadResult}{MriCloudR,character,character,logical}(object,
  jobId = "character", filename = tempfile(fileext = ".zip"),
  waitForJobToFinish = TRUE)

\S4method{downloadResult}{MriCloudR,character,missing,logical}(object,
  jobId = "character", filename = tempfile(fileext = ".zip"),
  waitForJobToFinish = TRUE)
}
\arguments{
\item{object}{Object of class \code{\link{MriCloudR-class}}.}

\item{jobId}{The jobId of the request, obtained from a processing request
such as \code{\link{t1Seg}}}

\item{filename}{Output filename for result.}

\item{waitForJobToFinish}{TRUE or FALSE.  If TRUE, \code{downloadResult}
will wait until the job is finished and then download the result (default
value is TRUE).  If FALSE, it will attempt to download the result but will
return if the job is not completed.}
}
\value{
TRUE if download successful. FALSE otherwise
}
\description{
\code{downloadResult} downloads the result of a processing request
associated with \code{jobId}, such as from a \code{\link{t1Seg}} request.
}
