% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MriCloudR.r
\docType{methods}
\name{dtiSeg,MriCloudR,DtiSegData-method}
\alias{dtiSeg,MriCloudR,DtiSegData-method}
\title{Submit a DTI segmentation request.}
\usage{
\S4method{dtiSeg}{MriCloudR,DtiSegData}(theObject, data = "DtiSegData")
}
\arguments{
\item{data}{A object of \code{\link{DtiSegData}} containing payload data and
metadata for the upload.}
}
\value{
Returns the job ID for the processing request.
}
\description{
\code{dtiMap} Submits an asynchronous DTI segmentation request, returning a
job ID as reference for subsequent requests to check and retreive results.
}

