% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DtiSegData.r
\docType{class}
\name{DtiSegData-class}
\alias{DtiSegData}
\alias{DtiSegData-class}
\title{DTI Data Payload Class}
\description{
This class is used to create an object payload for DTI mapping processing.
}
\section{Fields}{

\describe{
\item{\code{processingServer}}{processing server to use on backend}

\item{\code{dataZip}}{zip payload filename.  The zip file must contain the following files with indicated naming convention :
 1)    b0 image (subjectname_b0.img, subjectname_b0.hdr)
2)    Mean diffusion weighted images (subjectname_dwi.img, subjectname_dwi.hdr)
3)    Tensor file (subjectname_tensor.dat), which is a raw matrix with 6xD1xD2xD3)}

\item{\code{sliceType}}{"Axial"}

\item{\code{atlas}}{"Pediatric_168labels_12atlases_V1", "Adult_168labels_8atlases_V1". Defaults to "Adult_168labels_8atlases_V1".}

\item{\code{description}}{Any description of the dataset (OPTIONAL)}
}}

