% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MriCloudR.r
\docType{methods}
\name{isJobFinished,MriCloudR,character-method}
\alias{isJobFinished,MriCloudR,character-method}
\title{Check job status}
\usage{
\S4method{isJobFinished}{MriCloudR,character}(theObject, jobId = "character")
}
\arguments{
\item{jobId}{The jobId of the request, obtained from a processing request such as \code{\link{t1Seg}}}
}
\value{
\code{TRUE} if the job is completed, otherwise returns \code{FALSE}
}
\description{
\code{isJobFinished} checks status of processing for \code{jobId}.
}

