% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MriCloudR.r
\docType{methods}
\name{isJobFinished}
\alias{isJobFinished}
\alias{isJobFinished,MriCloudR,character-method}
\title{Check job status}
\usage{
isJobFinished(object, jobId = "character")

\S4method{isJobFinished}{MriCloudR,character}(object, jobId = "character")
}
\arguments{
\item{object}{Object of class \code{\link{MriCloudR-class}}.}

\item{jobId}{The jobId of the request, obtained from a processing request
such as \code{\link{T1SegData}}}
}
\value{
\code{TRUE} if the job is completed, otherwise returns \code{FALSE}
}
\description{
\code{isJobFinished} checks status of processing for \code{jobId}.
}
