% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MriCloudR.r
\docType{methods}
\name{listJobs}
\alias{listJobs}
\alias{listJobs,MriCloudR-method}
\title{List Jobs from MRI Cloud}
\usage{
listJobs(object)

\S4method{listJobs}{MriCloudR}(object)
}
\arguments{
\item{object}{Object of class \code{\link{MriCloudR-class}}.}
}
\value{
Returns the job ID for the processing request.
}
\description{
List Jobs from MRI Cloud
}
