% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MriCloudR.r
\docType{methods}
\name{login,MriCloudR,character,character-method}
\alias{login,MriCloudR,character,character-method}
\title{Login to MriCloud}
\usage{
\S4method{login}{MriCloudR,character,character}(theObject, username, password)
}
\arguments{
\item{username}{MriCloud username}

\item{password}{MriCloud password}
}
\description{
\code{login} logs into the MriCloud Api, which must be done before calling
any subsequent methods. If you do not have an MriCloud account, go to 
\url{https://braingps.mricloud.org} to register or retrieve forgotton 
credentials.
}

