% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MriCloudR.r
\docType{methods}
\name{login}
\alias{login}
\alias{login,MriCloudR,character,character-method}
\title{Login to MriCloud}
\usage{
login(object, username, password)

\S4method{login}{MriCloudR,character,character}(object, username, password)
}
\arguments{
\item{object}{Object of class \code{\link{MriCloudR-class}}.}

\item{username}{MriCloud username}

\item{password}{MriCloud password}
}
\description{
\code{login} logs into the MriCloud Api, which must be done before calling
any subsequent methods. If you do not have an MriCloud account, go to
\url{https://braingps.mricloud.org} to register or retrieve forgotton
credentials.
}
