% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MriCloudR.r
\docType{class}
\name{MriCloudR-class}
\alias{MriCloudR}
\alias{MriCloudR-class}
\title{A wrapper around the AnatomyWorks MriCloud Web API}
\description{
This class makes the MriCloud API functionality available in R,
encapsulating the http communications so that it behaves like a standard R
interface.
}
\section{Slots}{

\describe{
\item{\code{baseUrl}}{The root URL of the MRICloud API.  Default is \url{https://braingps.mricloud.org}.}

\item{\code{verbose}}{Verbose output}
}}

