% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MriCloudR.r
\docType{methods}
\name{t1Seg,MriCloudR,T1SegData-method}
\alias{t1Seg,MriCloudR,T1SegData-method}
\title{Submit a t1 segmentation request.}
\usage{
\S4method{t1Seg}{MriCloudR,T1SegData}(theObject, data = "T1SegData")
}
\arguments{
\item{data}{A object of \code{\link{T1SegData}} containing payload data and
metadata for the upload.}
}
\value{
Returns the job ID for the processing request.
}
\description{
\code{t1Seg} Submits an asynchronous t1 segmentation request, returning a
job ID as reference for subsequent requests to check and retreive results.
}

